Compute rand() challenges in unsigned arithmetic
SV_GetChallenge, SV_SpawnServer's checksum feed and CL_RequestMotd's
update server challenge build a 32-bit value as (rand() << 16) ^ rand().
RAND_MAX is 0x7fffffff on glibc and on Retro68 newlib, so the left shift
overflows a signed int whenever rand() returns 0x8000 or more. That is
undefined behaviour: UBSan's shift-base check reports it, and an
optimizer may assume it never happens.

Shift and xor the rand() results as unsigned, as ioquake3 does, and
convert the result back to int. On a 32-bit two's complement int the
bits are the ones the wrapped signed shift gave, so the challenge and
checksum feed values that reach the wire are unchanged.

The new challenge_rand runner stubs rand() and drives the real
SV_GetChallenge and CL_RequestMotd, plus the checksum feed statements
taken from sv_init.c, under -fsanitize=shift-base,signed-integer-overflow
with -fno-sanitize-recover, and checks each value against the wrapped
value computed in unsigned arithmetic.

Closes #415

+++ code/server/sv_client.c
@@ -73,7 +73,8 @@ void SV_GetChallenge( netadr_t from ) {
 		// this is the first time this client has asked for a challenge
 		challenge = &svs.challenges[oldest];
 
-		challenge->challenge = ( (rand() << 16) ^ rand() ) ^ svs.time;
+		challenge->challenge = (int)( ( (unsigned)rand() << 16 ) ^
+			(unsigned)rand() ^ (unsigned)svs.time );
 		challenge->adr = from;
 		challenge->firstTime = svs.time;
 		challenge->time = svs.time;
